Having thought more about it, I have decided it would be better to change the details concern into a controller that both frameworks can use. I think this works better and makes it more clear what is going on.
In the next PR I will update the old RM3830 procurement to use this controller too.
