+++ app/views/facilities_management/procurement_buildings/_buildings_and_services.html.erb
@@ -8,7 +8,7 @@
           </h2>
         </legend>
         <div id="assigning-services-hint" class="govuk-hint">
-          <%= t('.hint_text_html', services_link: link_to(t('.services_link_text'), facilities_management_rm6232_procurement_detail_path(procurement_id: @procurement.id, section: :services), class: 'govuk-link govuk-link--no-visited-state')) %>
+          <%= t('.hint_text_html', services_link: link_to(t('.services_link_text'), facilities_management_rm6232_procurement_procurement_detail_path(procurement_id: @procurement.id, section: :services), class: 'govuk-link govuk-link--no-visited-state')) %>
         </div>
         <%= displayed_error %>
         <div class="govuk-checkboxes buildings_and_services">
