+++ app/views/facilities_management/procurement_details/edit.html.erb
@@ -12,8 +12,8 @@
   </div>
 </div>
 
-<%= form_with url: facilities_management_rm6232_procurement_detail_path, model: @procurement, method: :put, local: 'false', html: { specialvalidation: true, novalidate: true } do |f| %>
-  <%= render partial: "facilities_management/shared/details/edit_partials/#{section}", locals: { f: f } %>
+<%= form_with url: facilities_management_rm6232_procurement_procurement_detail_path, model: @procurement, method: :put, local: 'false', html: { specialvalidation: true, novalidate: true } do |f| %>
+  <%= render partial: "facilities_management/procurement_details/edit_partials/#{section}", locals: { f: f } %>
 
   <div class="govuk-grid-row govuk-!-margin-top-5">
     <div class="govuk-grid-column-full">
