+++ app/views/facilities_management/procurement_details/edit_partials/_annual_contract_value.html.erb
