+++ app/views/facilities_management/procurement_details/edit_partials/_services.html.erb
