+++ app/views/facilities_management/procurement_details/edit_partials/contract_period/_call_off_extensions.html.erb
@@ -14,7 +14,7 @@
         <%= govuk_radios_conditional_area(:extensions_required, true) do %>
           <div id="radio-inner-content-for-call-off-extensions" class="radio-inner-content govuk-!-margin-top-3">
             <%= f.fields_for :call_off_extensions, call_off_extensions do |optional_period_field| %>
-              <%= render 'facilities_management/shared/details/edit_partials/contract_period/extension', f: optional_period_field %>
+              <%= render 'facilities_management/procurement_details/edit_partials/contract_period/extension', f: optional_period_field %>
             <% end %>
             <button id="add-contract-extension-button" class="govuk-button govuk-button--secondary govuk-!-margin-left-3 govuk-!-margin-top-3" aria-label="<%= t('.add_extension_period') %>">
               <%= t('.add_extension_period') %>
