+++ app/views/facilities_management/procurement_details/edit_partials/contract_period/_mobilisation_period.html.erb
