+++ app/views/facilities_management/procurement_details/kaminari/_first_page.html.erb
