+++ app/views/facilities_management/procurement_details/kaminari/_gap.html.erb
