+++ app/views/facilities_management/procurement_details/kaminari/_last_page.html.erb
