+++ app/views/facilities_management/procurement_details/kaminari/_next_page.html.erb
