+++ app/views/facilities_management/procurement_details/kaminari/_page.html.erb
