+++ app/views/facilities_management/procurement_details/kaminari/_paginator.html.erb
