+++ app/views/facilities_management/procurement_details/kaminari/_prev_page.html.erb
