+++ app/views/facilities_management/procurement_details/show.html.erb
@@ -11,7 +11,7 @@
   </div>
 </div>
 
-<%= render partial: "facilities_management/shared/details/show_partials/#{section}" %>
+<%= render partial: "facilities_management/procurement_details/show_partials/#{section}" %>
 
 <div class="govuk-grid-row govuk-!-margin-top-5">
   <div class="govuk-grid-column-full">
