+++ app/views/facilities_management/procurement_details/show_partials/_buildings_and_services.html.erb
