+++ app/views/facilities_management/procurement_details/show_partials/_contract_period.html.erb
