+++ app/views/facilities_management/procurement_details/show_partials/_services.html.erb
