--- app/views/facilities_management/shared/details/edit_partials/_contract_period.html.erb
@@ -1,13 +1,0 @@
-<div class="govuk-grid-row">
-  <div class="govuk-grid-column-full">
-    <div class="govuk-form-group">
-      <div class="contract-dates">
-        <%= render partial: 'facilities_management/shared/details/edit_partials/contract_period/initial_call_off_period', locals: { f: f } %>
-        <hr class="govuk-section-break govuk-section-break--m">
-        <%= render partial: 'facilities_management/shared/details/edit_partials/contract_period/mobilisation_period', locals: { f: f } %>
-        <hr class="govuk-section-break govuk-section-break--m">
-        <%= render partial: 'facilities_management/shared/details/edit_partials/contract_period/call_off_extensions', locals: { f: f } %>
-      </div>
-    </div>
-  </div>
-</div>
